rename Reader.c and Writer.c to avoid modules with the same name (and VS warnings)
