+++ modules/c/j2k/source/j2k_Writer.c
