+++ modules/c/nitf/source/NitfReader.c
