+++ modules/c/nitf/source/NitfWriter.c
